+++ projects/ngx-mask-lib/src/test/copy-paste.spec.ts
@@ -23,7 +23,7 @@ describe('Event: paste', () => {
 		component.mask = '00 - 0000 - 00000';
 		fixture.detectChanges();
 
-		const inputDebuggerElement = fixture.debugElement.query(By.css('#maska'));
+		const inputDebuggerElement = fixture.debugElement.query(By.css('#mask'));
 
 		const pasteData = new DataTransfer();
 		pasteData.setData('text', '123456789');
@@ -39,5 +39,27 @@ describe('Event: paste', () => {
 		expect(inputDebuggerElement.nativeElement.value).toBe('12 - 3456 - 789');
 
 		expect(inputDebuggerElement.nativeElement.selectionStart).toBe(15);
+	});
+	it('After paste to control cursor should be on the end of input for mask with separator', () => {
+		component.mask = 'separator.0';
+		component.thousandSeparator = ',';
+		fixture.detectChanges();
+
+		const inputDebuggerElement = fixture.debugElement.query(By.css('#mask'));
+
+		const pasteData = new DataTransfer();
+		pasteData.setData('text', '1234567');
+		inputDebuggerElement.triggerEventHandler('paste', pasteData);
+
+		inputDebuggerElement.nativeElement.value = pasteData.getData('text/plain');
+		inputDebuggerElement.triggerEventHandler('input', {
+			target: inputDebuggerElement.nativeElement,
+		});
+
+		fixture.detectChanges();
+
+		expect(inputDebuggerElement.nativeElement.value).toBe('1,234,567');
+
+		expect(inputDebuggerElement.nativeElement.selectionStart).toBe(9);
 	});
 });